feat(be): update Ticket and TicketService to use BigInteger for blockchainId and add ticket invalidation logic

+++ backend/src/main/java/com/ticketblock/entity/Ticket.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @AllArgsConstructor
@@ -21,7 +22,7 @@ public class Ticket {
     @EqualsAndHashCode.Include
     private Integer id;
 
-    private Integer blockchainId;
+    private BigInteger blockchainId;
 
     @ManyToOne()
     @JoinColumn(name = "event_id")
